+++ actions/aws-s3/index.js
@@ -10,7 +10,73 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const utils = require('./utils')
+const utils = require('tvm-utils')
+const aws = require('aws-sdk')
+const joi = require('@hapi/joi')
+
+// Specific utils
+/**
+ * Reads ./policy.json and replaces __RESOURCE_NAME__ and __USER_NAMESPACE__ placeholders
+ * with parameters
+ * @param  {string} resourceName
+ * @param  {string} namespace
+ * @returns {string}
+ */
+function generatePolicy (resourceName, namespace) {
+  return JSON.stringify(require('./policy.json'))
+    .replace(/__RESOURCE_NAME__/g, resourceName)
+    .replace(/__USER_NAMESPACE__/g, namespace)
+}
+
+/**
+ * @param  {string} accessKeyId aws key id of account allowed to call sts:GetFederationToken
+ * @param  {string} secretAccessKey aws secret key
+ * @param  {string} policy for restricting tmp tokens
+ * @param  {string} username name of federated user for tokens
+ * @param  {number} expiryDuration validity duration in seconds
+ * @returns {Promise<object>} {accessKeyId, secretAccessKey, sessionToken, expiration}
+ * @throws on error
+ */
+async function generateTmpAWSToken (accessKeyId, secretAccessKey, policy, username, expiryDuration) {
+  const sts = new aws.STS({
+    accessKeyId,
+    secretAccessKey,
+    apiVersion: '2011-06-15'
+  })
+  // When calling aws, for some errors (e.g undefined accessKeyId) the action
+  // seems to hang when running in OpenWhisk. A temporary hack is to enforce a
+  // timeout. An error response is correctly returned from AWS when running the
+  // action locally. Todo: investigate why this doesn't work in prod
+  const stsTimeout = 20000
+  let timeoutId
+  let stsRes
+  try {
+    stsRes = await Promise.race([
+      sts.getFederationToken({
+        DurationSeconds: expiryDuration,
+        Policy: policy,
+        Name: username
+      }).promise().catch(e => { throw new Error(`AWS STS Error: ${e.message}`) }),
+      new Promise((resolve, reject) => {
+        timeoutId = setTimeout(() => reject(new Error(`sts.getFederationToken timed out after ${stsTimeout}`)), stsTimeout)
+      })
+    ])
+  } finally {
+    clearTimeout(timeoutId)
+  }
+  // return error instead of throw like validation functions ?
+  // in that case try/catch sts call
+  if (!stsRes.Credentials) {
+    throw new Error(`sts.getFederationToken didn't return with Credentials, instead: [${JSON.stringify(stsRes)}]`)
+  }
+
+  return {
+    accessKeyId: stsRes.Credentials.AccessKeyId,
+    secretAccessKey: stsRes.Credentials.SecretAccessKey,
+    sessionToken: stsRes.Credentials.SessionToken,
+    expiration: stsRes.Credentials.Expiration
+  }
+}
 
 /**
  * @param  {string} err - error message
@@ -26,12 +92,12 @@ function errorResponse (err, status) {
 /**
  * @param {object} params - the input params
  *
- * @params {string} params.s3Bucket - privided by owner, default final
- * @params {number} params.expiryDuration - privided by owner, default final
- * @params {string} params.whitelist - privided by owner, default final
- * @params {string} params.awsAccessKeyId - privided by owner, default final
- * @params {string} params.awsSecretAccessKey - privided by owner, default final
- * @params {string} params.owApiHost - privided by owner, default final
+ * @params {string} params.s3Bucket - provided by owner, default final
+ * @params {number} params.expiryDuration - provided by owner, default final
+ * @params {string} params.whitelist - provided by owner, default final
+ * @params {string} params.awsAccessKeyId - provided by owner, default final
+ * @params {string} params.awsSecretAccessKey - provided by owner, default final
+ * @params {string} params.owApiHost - provided by owner, default final
  *
  * @params {string} params.owAuth - user's OpenWhisk Basic Token
  * @params {string} params.owNamespace - user's OpenWhisk Namespace
@@ -42,11 +108,13 @@ function errorResponse (err, status) {
 async function main (params) {
   try {
     // 0. validate params
-    const resParams = utils.validateParams(params)
+    const resParams = utils.validateParams(params, { s3Bucket: joi.string().required(), awsAccessKeyId: joi.string().required(), awsSecretAccessKey: joi.string().required() })
     if (resParams.error) {
       console.warn(`Bad request: ${resParams.error.message}`)
       return errorResponse(`${resParams.error.message}`, 400)
     }
+    // important !! as joi accepts '123'
+    params.expiryDuration = parseInt(params.expiryDuration)
 
     console.log(`Incoming request for [ ${params.owNamespace}, ${params.owAuth.split(':')[0]} ]`)
 
@@ -66,11 +134,11 @@ async function main (params) {
     console.log('Request is authorized')
 
     // 3. Get AWS tokens
-    const policy = utils.generatePolicy(params.s3Bucket, params.owNamespace)
+    const policy = generatePolicy(params.s3Bucket, params.owNamespace)
 
     console.log('Policy generated')
 
-    const tmpCreds = await utils.generateTmpAWSToken(params.awsAccessKeyId, params.awsSecretAccessKey, policy, params.owNamespace, params.expiryDuration)
+    const tmpCreds = await generateTmpAWSToken(params.awsAccessKeyId, params.awsSecretAccessKey, policy, params.owNamespace, params.expiryDuration)
 
     console.log(`AWS tokens generated`)
     console.log(`End of request`)
