+++ lib/tvm-utils/index.js
@@ -10,15 +10,35 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 const openwhisk = require('openwhisk')
+const joi = require('@hapi/joi')
 
-// TODO MODULARIZE UTILS FROM OTHER ACTIONS
+/**
+ * @param {object} params
+ * @returns {null | object} if validation failed returns error in res.error
+ */
+function validateParams (params, subSchema) {
+  const schema = joi.object().label('params').keys({
+    // default params
+    // must be final params, especially the whitelist, s3Bucket, containerName and expiryDuration for security reasons
+    ...subSchema,
+    /// common params
+    expiryDuration: joi.number().required(),
+    whitelist: joi.string().required(),
+    owApihost: joi.string().uri().required(),
+    // those are user openwhisk credentials passed as request params
+    owAuth: joi.string().required(),
+    owNamespace: joi.string().required()
+  }).pattern(/^$/, joi.any()).pattern(/^__ow_.+$/, joi.any()) // this means: allow all unknown parameters that start with __ow_ and ''
+  // somehow the api-gateway adds a '' field to the params if no path was provided, so we allow it
+  return joi.validate(params, schema)
+}
 // generic helpers
 
 /**
  * @param  {string} apihost
  * @param  {string} namespace
  * @param  {string} auth
- * @returns {null | object} if validation failed returns error in res.error
+ * @returns {Promise<null | object>} if validation failed returns error in res.error
  */
 async function validateOWCreds (apihost, namespace, auth) {
   const ow = openwhisk({ api_key: auth, apihost: apihost })
@@ -47,27 +67,11 @@ async function validateOWCreds (apihost, namespace, auth) {
  */
 function isWhitelisted (namespace, whitelist) {
   if (whitelist.trim() === '*') return true
-  whitelist = whitelist.split(',').map(ns => ns.trim())
-  return whitelist.includes(namespace)
-}
-
-/**
- * creates a container if not exists
- *
- * @param {azure.ContainerURL} containerURL azure ContainerUrl
- * @param {azure.Aborter} aborter azure Aborter
- * @param {object} [options] azure container creation options
- */
-async function createContainerIfNotExists (containerURL, aborter, options) {
-  try {
-    await containerURL.create(aborter, options)
-  } catch (e) {
-    if (e.body.Code !== 'ContainerAlreadyExists') throw e
-  }
+  return whitelist.split(',').map(ns => ns.trim()).includes(namespace)
 }
 
 module.exports = {
-  isWhitelisted,
-  validateOWCreds,
-  createContainerIfNotExists
+  isWhitelisted: isWhitelisted,
+  validateParams: validateParams,
+  validateOWCreds: validateOWCreds
 }
