+++ lib/tvm-utils/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "tvm-utils",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@hapi/joi": "^15.1.0",
+    "openwhisk": "^3.19.0"
+  }
+}
